# Grp fill perf