--- a/pandas/_libs/groupby_helper.pxi.in
+++ b/pandas/_libs/groupby_helper.pxi.in
@@ -987,7 +987,7 @@ def group_fillna_indexer(ndarray[int64_t] out,
 
     N = len(out)
 
-    sorted_labels = np.argsort(labels)
+    sorted_labels = np.argsort(labels).view(dtype=np.int64)
     if method == 'bfill':
         sorted_labels = sorted_labels[::-1]