--- a/pandas/_libs/groupby_helper.pxi.in
+++ b/pandas/_libs/groupby_helper.pxi.in
@@ -906,7 +906,7 @@ def group_cumsum(numeric[:, :] out,
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
-def group_shift_indexer(int64_t[:] out, int64_t[:] labels,
+def group_shift_indexer(ndarray[int64_t] out, ndarray[int64_t] labels,
                         int ngroups, int periods):
     cdef:
         Py_ssize_t N, i, j, ii
@@ -957,21 +957,19 @@ def group_shift_indexer(int64_t[:] out, int64_t[:] labels,
 
 @cython.wraparound(False)
 @cython.boundscheck(False)
-def group_fillna_indexer(ndarray[int64_t] out,
-                         ndarray[uint8_t] mask,
-                         ndarray[int64_t] labels,
-                         object method,
+def group_fillna_indexer(ndarray[int64_t] out, ndarray[int64_t] labels,
+                         ndarray[uint8_t] mask, object direction,
                          int64_t limit):
-    """Fills values forwards or backwards within a group
+    """Indexes how to fill values forwards or backwards within a group
 
     Parameters
     ----------
     out : array of int64_t values which this method will write its results to
         Missing values will be written to with a value of -1
-    mask : array of int64_t values where a 1 indicates a missing value
     labels : array containing unique label for each group, with its ordering
         matching up to the corresponding record in `values`
-    method : {'ffill', 'bfill'}
+    mask : array of int64_t values where a 1 indicates a missing value
+    direction : {'ffill', 'bfill'}
         Direction for fill to be applied (forwards or backwards, respectively)
     limit : Consecutive values to fill before stopping, or -1 for no limit
 
@@ -987,8 +985,11 @@ def group_fillna_indexer(ndarray[int64_t] out,
 
     N = len(out)
 
+    # Make sure all arrays are the same size
+    assert N == len(labels) == len(mask)
+
     sorted_labels = np.argsort(labels).astype(np.int64, copy=False)
-    if method == 'bfill':
+    if direction == 'bfill':
         sorted_labels = sorted_labels[::-1]
 
     with nogil:
@@ -1004,6 +1005,7 @@ def group_fillna_indexer(ndarray[int64_t] out,
                 curr_fill_idx = idx
 
             out[idx] = curr_fill_idx
+
             # If we move to the next group, reset
             # the fill_idx and counter
             if i == N - 1 or labels[idx] != labels[sorted_labels[i+1]]: