--- a/pandas/_libs/groupby_helper.pxi.in
+++ b/pandas/_libs/groupby_helper.pxi.in
@@ -907,15 +907,13 @@ def group_cumsum(numeric[:, :] out,
 @cython.boundscheck(False)
 @cython.wraparound(False)
 def group_shift_indexer(ndarray[int64_t] out, ndarray[int64_t] labels,
-                        int64_t ngroups, **kwargs):
+                        int ngroups, int periods):
     cdef:
         Py_ssize_t N, i, j, ii
-        int offset, sign, periods
+        int offset, sign
         int64_t lab, idxer, idxer_slot
         int64_t[:] label_seen = np.zeros(ngroups, dtype=np.int64)
         int64_t[:, :] label_indexer
-
-    periods = kwargs['periods']
 
     N, = (<object> labels).shape
 
@@ -960,7 +958,8 @@ def group_shift_indexer(ndarray[int64_t] out, ndarray[int64_t] labels,
 @cython.wraparound(False)
 @cython.boundscheck(False)
 def group_fillna_indexer(ndarray[int64_t] out, ndarray[int64_t] labels,
-                         ndarray[uint8_t] mask, **kwargs):
+                         ndarray[uint8_t] mask, object direction,
+                         int64_t limit):
     """Indexes how to fill values forwards or backwards within a group
 
     Parameters
@@ -981,10 +980,8 @@ def group_fillna_indexer(ndarray[int64_t] out, ndarray[int64_t] labels,
     cdef:
         Py_ssize_t i, N
         ndarray[int64_t] sorted_labels
-        int64_t limit, idx, curr_fill_idx=-1, filled_vals=0
+        int64_t idx, curr_fill_idx=-1, filled_vals=0
 
-    direction = kwargs['direction']
-    limit = kwargs['limit']
     N = len(out)
 
     # Make sure all arrays are the same size