--- a/pandas/_libs/groupby_helper.pxi.in
+++ b/pandas/_libs/groupby_helper.pxi.in
@@ -907,13 +907,15 @@ def group_cumsum(numeric[:, :] out,
 @cython.boundscheck(False)
 @cython.wraparound(False)
 def group_shift_indexer(ndarray[int64_t] out, ndarray[int64_t] labels,
-                        int ngroups, int periods):
+                        int64_t ngroups, **kwargs):
     cdef:
         Py_ssize_t N, i, j, ii
-        int offset, sign
+        int offset, sign, periods
         int64_t lab, idxer, idxer_slot
         int64_t[:] label_seen = np.zeros(ngroups, dtype=np.int64)
         int64_t[:, :] label_indexer
+
+    periods = kwargs['periods']
 
     N, = (<object> labels).shape
 
@@ -958,8 +960,7 @@ def group_shift_indexer(ndarray[int64_t] out, ndarray[int64_t] labels,
 @cython.wraparound(False)
 @cython.boundscheck(False)
 def group_fillna_indexer(ndarray[int64_t] out, ndarray[int64_t] labels,
-                         ndarray[uint8_t] mask, object direction,
-                         int64_t limit):
+                         ndarray[uint8_t] mask, **kwargs):
     """Indexes how to fill values forwards or backwards within a group
 
     Parameters
@@ -980,9 +981,10 @@ def group_fillna_indexer(ndarray[int64_t] out, ndarray[int64_t] labels,
     cdef:
         Py_ssize_t i, N
         ndarray[int64_t] sorted_labels
-        int64_t curr_fill_idx=-1
-        int64_t idx, filled_vals=0
+        int64_t limit, idx, curr_fill_idx=-1, filled_vals=0
 
+    direction = kwargs['direction']
+    limit = kwargs['limit']
     N = len(out)
 
     # Make sure all arrays are the same size