… asv_bench/benchmarks/groupby.py
… doc/source/whatsnew/v0.23.0.txt
… pandas/tests/groupby/test_groupby.py