- [X] closes #11296
- [X] tests added / passed
- [X] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [ ] whatsnew entry

I am not a fan of how I've implemented this in `groupby.py` but I think this change highlights even more the need for some refactoring of that module, specifically in how Cython transformations are getting dispatched. This still works in the meantime but open to any feedback on how the methods are getting wired back to the Cython layer.

Below are ASVs for the change

```bash
       before           after         ratio
     [d9551c8e]       [5e007f86]
+      81.6±0.3μs         96.7±2μs     1.19  groupby.GroupByMethods.time_method('float', 'count')
+         455±3μs         516±20μs     1.13  groupby.GroupByMethods.time_method('float', 'cummin')
+       306±0.6μs          340±5μs     1.11  groupby.GroupByMethods.time_method('float', 'prod')
-       142±0.4ms          280±6μs     0.00  groupby.GroupByMethods.time_method('int', 'bfill')
-        230±10ms          327±7μs     0.00  groupby.GroupByMethods.time_method('float', 'bfill')
-       135±0.8ms        189±0.3μs     0.00  groupby.GroupByMethods.time_method('int', 'ffill')
-         227±2ms        184±0.6μs     0.00  groupby.GroupByMethods.time_method('float', 'ffill')
```