@chris-b1 2018-02-13
One possibility to clean this up would be to replicate `GroupBy.shift`
https://github.com/pandas-dev/pandas/blob/07137a5acdfc761c7cc30a081600e4c4f23c7d3d/pandas/core/groupby.py#L1846

Rather than type specific functions, there the cython routine computes a single indexer, then the normal `take` functions are used to make the actual output.
https://github.com/pandas-dev/pandas/blob/07137a5acdfc761c7cc30a081600e4c4f23c7d3d/pandas/_libs/groupby_helper.pxi.in#L991

Probably a bit slower than what you do here, but would be surprised if it's bad.