@WillAyd 2018-02-13
Thanks for the idea. I think I could make something like that work here and that would definitely simplify things