@WillAyd 2018-02-13
Just simplified the code to mirror what we do for `shift`. The `groupby.py` changes are MUCH simpler in that case. There doesn't appear to be any significant change to benchmarks either, with updated results below:

```bash
       before           after         ratio
     [07137a5a]       [0939aeae]
+           983ms            1.23s     1.25  groupby.GroupByMethods.time_method('float', 'pct_change')
+         356±4μs         443±20μs     1.24  groupby.GroupByMethods.time_method('float', 'prod')
+      84.6±0.2μs          101±3μs     1.19  groupby.GroupByMethods.time_method('int', 'count')
+         185±1μs          218±3μs     1.18  groupby.GroupByMethods.time_method('int', 'cumcount')
+         161±3μs          185±3μs     1.15  groupby.GroupByMethods.time_method('float', 'cumcount')
+         334±9μs         382±20μs     1.14  groupby.GroupByMethods.time_method('float', 'median')
+         461±3μs         523±10μs     1.13  groupby.GroupByMethods.time_method('int', 'cummax')
+       344±0.8μs         387±10μs     1.12  groupby.GroupByMethods.time_method('float', 'nunique')
+      64.5±0.2μs       72.1±0.9μs     1.12  groupby.GroupByMethods.time_method('int', 'size')
+         456±2μs         504±20μs     1.11  groupby.GroupByMethods.time_method('float', 'cummin')
+         133±5ms          146±3ms     1.10  groupby.GroupByMethods.time_method('int', 'any')
-        769±20μs         684±30μs     0.89  groupby.GroupByMethods.time_method('int', 'cumprod')
-       893±100μs         598±30μs     0.67  groupby.GroupByMethods.time_method('int', 'sem')
-         145±2ms          275±7μs     0.00  groupby.GroupByMethods.time_method('int', 'bfill')
-         154±3ms          245±1μs     0.00  groupby.GroupByMethods.time_method('int', 'ffill')
-         228±5ms         270±10μs     0.00  groupby.GroupByMethods.time_method('float', 'bfill')
-         237±5ms        232±0.5μs     0.00  groupby.GroupByMethods.time_method('float', 'ffill')
```