@WillAyd 2018-02-19
Latest commit includes a shared function between `fill` and `shift`, which can theoretically house `any` / `all` implementations with a little clean up as well. As touched on in previous conversations, if we expand this it should probably be in its own module

Open to review and plan to change the location of the func within the module / update docs, but I wanted to push because this commit will cause tests to fail at the following location:

https://github.com/pandas-dev/pandas/blob/718d0679600fa025872d8b66c198a6781b312d70/pandas/tests/groupby/test_transform.py#L555

The reason this is failing is because I replaced `output = {}` in the original `shift` code with `output = collections.OrderedDict()` in the newly shared function. I'm not entirely clear on why that would impact the test case, but it seems like a more explicit code path regardless. 

I believe the test case is wrong anyway and the failing line should either have a `.sort_index` call or be placed before the `expected = expected.sort_index(axis=1)` call a few lines ahead of it, but wanted to get your input and make sure I am not misreading that