@WillAyd 2018-02-19
To clarify my point about having a separate module for `_get_cythonized_result` I don't think that method as a stand-alone needs it's own module, but thinking long term and going back to the discussion in #19481 a class to more effectively dispatch all of the functions (agg and transform included) may make sense as a sub-module of `groupby.py` and `_get_cythonized_result` could potentially evolve into that