@pep8speaks 2018-02-19
Hello @WillAyd! Thanks for updating the PR.



 - In the file [`pandas/core/groupby.py`](https://github.com/pandas-dev/pandas/blob/0068948591080c9368250b70b7b5d0b4eed95cd1/pandas/core/groupby.py), following are the PEP8 issues :

> [Line 4640:14](https://github.com/pandas-dev/pandas/blob/0068948591080c9368250b70b7b5d0b4eed95cd1/pandas/core/groupby.py#L4640): [E131](https://duckduckgo.com/?q=pep8%20E131) continuation line unaligned for hanging indent



 - In the file [`pandas/tests/groupby/test_transform.py`](https://github.com/pandas-dev/pandas/blob/0068948591080c9368250b70b7b5d0b4eed95cd1/pandas/tests/groupby/test_transform.py), following are the PEP8 issues :

> [Line 557:17](https://github.com/pandas-dev/pandas/blob/0068948591080c9368250b70b7b5d0b4eed95cd1/pandas/tests/groupby/test_transform.py#L557): [E124](https://duckduckgo.com/?q=pep8%20E124) closing bracket does not match visual indentation