@jreback 2018-02-22
ping when switch to concat and green