@WillAyd 2018-02-22
@jreback all green, though it looks like AppVeyor was never triggered