@WillAyd 2018-02-24
Travis failure on latest commit looks unrelated (something with pyarrow?)